Update Sony - PlayStation 2.json
About "Attack of the Toybots":
`Nicktoons - Attack of the Toybots` is an english only release for the PAL zone (both EU and AUS) but `Nickelodeon SpongeBob and Friends - Attack of the Toybots` is a Multi6 release for Europe (no Australia but it's exactly the same game, minus the title screen). Retool only takes the `Nicktoons` instead of the other one with the previous setting (treating `Nickelodeon SpongeBob and Friends - Attack of the Toybots` as a regional equivalent)... What I'm trying to do here, by putting 0 in the line of `Nickelodeon SpongeBob and Friends - Attack of the Toybots` is to give this release the priority because any PAL player can give this one a go instead of `Nicktoons` which is english only.